Refactor sync logic and update Syncable protocol
Moved synchronization logic from Sync.swift to a new SyncDriver.swift, removing SyncError.swift and Syncable+Shuffling.swift. Updated the Syncable protocol to use associated types for value, and adjusted SyncGroup and SyncableConformance implementations for type safety and clarity. This refactor improves modularity and prepares for more flexible batch processing.
