+++ Sources/Scout/Core/SyncGroup.swift
@@ -8,17 +8,13 @@
 import CloudKit
 import CoreData
 
-typealias SyncValue = MatrixValue & CKRecordValueProtocol & AdditiveArithmetic & Sendable
-
 struct SyncGroup<T: SyncValue>: @unchecked Sendable {
     let recordType: String
     let name: String
     let date: Date
-    let objects: [Syncable]
+    let objects: [any Syncable]
     let fields: [String: T]
-}
 
-extension SyncGroup {
     func newMatrix() -> Matrix<Cell<T>> {
         Matrix(
             date: date,
