+++ Sources/Scout/Core/Syncable.swift
@@ -8,30 +8,14 @@
 import CloudKit
 import CoreData
 
-/// Core Data models that can be synchronized as small, logical batches.
-///
-/// Contract:
-/// - `group(in:)` returns **one** batch (or `nil` if nothing pending).
-/// - Implementations are free to choose the grouping key (e.g. week/name).
-/// - Keep batches small (use a seed row + its key to collect the set).
-///
+typealias SyncValue = MatrixValue & CKRecordValueProtocol & AdditiveArithmetic & Sendable
+
 protocol Syncable: NSManagedObject {
-
-    /// Returns a batch of currently-unsynced objects, or `nil` if none.
-    ///
-    /// Implementations should:
-    /// - use one “seed” unsynced row to determine the batch key,
-    /// - fetch the rest of the unsynced rows matching that key,
-    /// - map them into `SyncGroup`.
-    ///
-    static func group(in context: NSManagedObjectContext) throws -> SyncGroup?
-
-    /// Whether this instance has been sent upstream.
+    associatedtype Value: SyncValue
+    static func group(in context: NSManagedObjectContext) throws -> SyncGroup<Value>?
     var isSynced: Bool { get set }
 }
 
-/// Errors for missing required fields when building a batch.
-///
 enum SyncableError: Error {
     case missingProperty(String)
